Make new gcc -Wall happy

+++ Modules/_tkinter.c
@@ -358,7 +358,7 @@ Tkapp_New(screenName, baseName, className, interactive)
 	}
 
 	strcpy(argv0, className);
-	if (isupper(argv0[0]))
+	if (isupper((int)(argv0[0])))
 		argv0[0] = tolower(argv0[0]);
 	Tcl_SetVar(v->interp, "argv0", argv0, TCL_GLOBAL_ONLY);
 	ckfree(argv0);
